refactor(routes): extract jellyfin audio subtitle routes
